feat(#2146): fix all qulice suggestions
